commented java code, added some safety + code re-use

+++ src/main/java/com/nitro/clients/kafka/TestUtils.java
@@ -7,13 +7,17 @@ import java.net.ServerSocket;
 import java.util.Random;
 
 class TestUtils {
+
     private static final Random RANDOM = new Random();
 
     private TestUtils() {
     }
 
-    public static File constructTempDir(String dirPrefix) {
-        File file = new File(System.getProperty("java.io.tmpdir"), dirPrefix + RANDOM.nextInt(10000000));
+    public static File constructTempDir(final String dirPrefix) {
+        final File file = new File(
+                System.getProperty("java.io.tmpdir"),
+                dirPrefix + RANDOM.nextInt(10000000)
+        );
         if (!file.mkdirs()) {
             throw new RuntimeException("could not create temp directory: " + file.getAbsolutePath());
         }
@@ -21,27 +25,17 @@ class TestUtils {
         return file;
     }
 
-    public static int getAvailablePort() {
-        try {
-            ServerSocket socket = new ServerSocket(0);
-            try {
-                return socket.getLocalPort();
-            } finally {
-                socket.close();
-            }
-        } catch (IOException e) {
-            throw new IllegalStateException("Cannot find available port: " + e.getMessage(), e);
-        }
-    }
-
-    public static boolean deleteFile(File path) throws FileNotFoundException {
+    public static boolean deleteFile(final File path) throws FileNotFoundException {
         if (!path.exists()) {
             throw new FileNotFoundException(path.getAbsolutePath());
         }
         boolean ret = true;
         if (path.isDirectory()) {
-            for (File f : path.listFiles()) {
-                ret = ret && deleteFile(f);
+            final File[] fis = path.listFiles();
+            if(fis != null){
+                for (final File f : fis) {
+                    ret = ret && deleteFile(f);
+                }
             }
         }
         return ret && path.delete();
